Infrastructure/visualization/requirements.txt
influxdb-client[ciso]
dash
polars
dash-leaflet
numpy
pandas
pyarrow
kafka-python